+++ changelog.md
@@ -2,6 +2,7 @@
 
 ## Unreleased
 
+* Added `from_i8x32` and `From<i8x32>` to `i16x32`.
 * Added runtime-index shuffle functions. Previous functions `swizzle` and
   `swizzle_relaxed` have been deprecated. See
   [PR #317](https://github.com/lokathor/wide/pull/317) for more information.